Refactor MS/MS reading from the TDF tables.
Refactored all of .do_calculate_core_ms2_information because old implementation was very slow. Now we use fast dplyr table joins to extract all relevant infomation from the tables. There's still room for future improvement, probably save the tables in the backendobject when initializing and also filtering them when subsetting the backend. In this way, you could keep the table sizes as small as needed and not have to read the long tables from each file if the backend is only storing a couple spectra.

Added depencencies on dplyr (mandatory for fast table joins) and utils (to automatically fill the `version` slot when creating a new backend object)

Temptative bump to 0.2.0 due to the addition of methods: precursorMz, precursorCharge, etc.